requirements-dev.txt
flake8==4.0.1
mypy==0.931
loguru-mypy==0.0.4
pytest==6.2.5
pytest_check==1.0.4
sphinx-panels==0.6.0
sphinx-rtd-theme==1.0.0
pywin32-stubs==0.1.4; sys_platform == "win32"